Support running on Linux, pointed at a Windows MWO installation.
java.io.File.equals has operating system-specific behavior: on Linux,
it is case sensitive, while on Windows, it is not. The Ice Ferret has a
pak file named "Iceferret.pak", so lookups of "iceferret.pak" succeed
on Windows but not Linux. Fix this by manually lowercasing the paths in
our lookup table.